requirements.txt
discord.py
fastapi
uvicorn[standard]
python-dotenv
